fix to match to foodcritic rules at part that does not modify the chefspec

+++ site-cookbooks/backup_restore/recipes/configure_postgresql.rb
@@ -28,16 +28,15 @@ backup_model :postgresql do
 end
 
 # set proxy environment if use_proxy
-if node['backup_restore']['config']['use_proxy']
-  ruby_block 'set_proxy_env' do
-    block do
-      proxy_url = "http://#{node['backup_restore']['config']['proxy_host']}:#{node['backup_restore']['config']['proxy_port']}/"
-      cron_file = '/etc/cron.d/postgresql_backup'
-      file = Chef::Util::FileEdit.new(cron_file)
-      file.insert_line_after_match(/# Crontab for/, "https_proxy=#{proxy_url}")
-      file.insert_line_after_match(/# Crontab for/, "http_proxy=#{proxy_url}")
-      file.write_file
-      File.delete("#{cron_file}.old") if File.exist?("#{cron_file}.old")
-    end
+ruby_block 'set_proxy_env' do
+  block do
+    proxy_url = "http://#{node['backup_restore']['config']['proxy_host']}:#{node['backup_restore']['config']['proxy_port']}/"
+    cron_file = '/etc/cron.d/postgresql_backup'
+    file = Chef::Util::FileEdit.new(cron_file)
+    file.insert_line_after_match(/# Crontab for/, "https_proxy=#{proxy_url}")
+    file.insert_line_after_match(/# Crontab for/, "http_proxy=#{proxy_url}")
+    file.write_file
+    File.delete("#{cron_file}.old") if File.exist?("#{cron_file}.old")
   end
+  only_if node['backup_restore']['config']['use_proxy']
 end
